requirements.txt
Flask==3.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
PyMySQL==1.1.1
pypng==0.20220715.0
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==7.4.2
requests==2.32.0
schedule==1.2.2
gunicorn==21.2.0
typing_extensions==4.11.0
urllib3==2.2.2
Werkzeug==3.1.3
cryptography>=42.0.0
Pillow==10.2.0
Flask-Limiter==3.5.1
Flask-WTF==1.2.1
PyJWT==2.8.0
